[development-plastic-transaction] feat: add points transfer and listing to users

+++ core/domain/src/main/java/com/clastic/domain/repository/PlasticTransactionRepository.kt
@@ -9,7 +9,7 @@ interface PlasticTransactionRepository {
         date: Date,
         dropPointId: String,
         ownerId: String,
-        totalPoints: Int,
+        totalPoints: Long,
         userId: String,
         plasticTransactionItemList: List<PlasticTransactionItem>,
         onPostSuccess: (transactionId: String) -> Unit,
